Actualizados los nombres de algunos punteros para quitar la notación hungara y añadidos algunos includes faltantes. Añadidos varios casts en operaciones y actualizado el tipo de algún atributo para evitar casts. Actualizado el vcxproj a warnings Level 4

+++ Game/Map/Tile.cpp
@@ -17,7 +17,7 @@ void Tile::attatchToGrid(Grid* grid, bool isBorderTile) {
     if (isBorderTile)
         sfSprite.setColor(sf::Color::Black);    // TODO cambiar esto para que muestre otro tile en vez de colorearlo (cosa que no hace).
 
-    sfSprite.setPosition(_grid->_posX * 32, _grid->_posY * 32);
+    sfSprite.setPosition((float)(_grid->_posX * 32), (float)(_grid->_posY * 32));
 }
 
 void Tile::detatchFromGrid() {
